+++ packages/compass-proto/src/components/RightSidebar/index.ts
@@ -1,0 +1,2 @@
+export { default as RightSidebar } from './RightSidebar';
+export type { RightSidebarProps } from './RightSidebar';