+++ src/guidelines/_components/ChannelHeaderGuidelineAnatomy.tsx
@@ -1,4 +1,4 @@
-import { ChannelHeader } from '@mattermost/compass-ui/components/channel-header';
+import { ChannelHeader } from '@mattermost/compass-proto';
 import AnatomyStage from '@/guidelines/_components/AnatomyStage';
 import styles from './ChannelHeaderGuidelineAnatomy.module.scss';