+++ src/guidelines/layouts/mobile-home/mobile-home.specimen.tsx
@@ -1,7 +1,5 @@
 import { useState} from 'react';
-import { MobileChannelsSidebar } from '@mattermost/compass-proto';
-import { type MobileTabBarTab } from '@mattermost/compass-proto';
-import { MobileHome, MobileTabBar, MobileTeamSidebar } from '@mattermost/compass-proto';
+import { MobileChannelsSidebar, type MobileTabBarTab, MobileHome, MobileTabBar, MobileTeamSidebar } from '@mattermost/compass-proto';
 import DeviceFrame from '@/components/layout/DeviceFrame';
 import avatarAikoTan from '@/assets/avatars/Aiko Tan.png';
 import avatarArjunPatel from '@/assets/avatars/Arjun Patel.png';