+++ src/guidelines/patterns/message/message.specimen.tsx
@@ -1,5 +1,5 @@
-import { Message } from '@mattermost/compass-ui/components/message';
-import { messageStyles } from '@mattermost/compass-ui/components/message';
+import { Message } from '@mattermost/compass-proto';
+import { messageStyles } from '@mattermost/compass-proto';
 import { Divider } from '@mattermost/compass-ui/components/divider';
 import avatarLeonard from '@/assets/avatars/Leonard Riley.png';
 import avatarDanielle from '@/assets/avatars/Danielle Okoro.png';